Move numpy interop defines into a separate header from numpy interop. Also remove unnecessary do_import_array.hpp and import_numpy call in config.pyx.
