+++ dynd/config.pyx
@@ -8,12 +8,6 @@ cdef extern from 'exception_translation.hpp':
 cdef void translate_exception():
     _translate_exception()
 
-cdef extern from 'do_import_array.hpp':
-    pass
-
-cdef extern from 'numpy_interop.hpp' namespace 'pydynd':
-    void import_numpy()
-
 cdef extern from 'git_version.hpp' namespace 'pydynd':
     extern char[] dynd_python_version_string
     extern char[] dynd_python_git_sha1
@@ -25,9 +19,6 @@ _dynd_version_string = str(<char *>dynd_version_string)
 _dynd_git_sha1 = str(<char *>dynd_git_sha1)
 _dynd_python_version_string = str(<char *>dynd_python_version_string)
 _dynd_python_git_sha1 = str(<char *>dynd_python_git_sha1)
-
-# Initialize Numpy
-import_numpy()
 
 # Exceptions to convert from C++
 class BroadcastError(Exception):
