--- dynd/include/do_import_array.hpp
@@ -1,2 +1,0 @@
-// Trigger the array import (inversion of NO_IMPORT_ARRAY)
-#define NUMPY_IMPORT_ARRAY
